Move classes to correct package
