+++ runners/maven/src/main/scala/stryker4s/maven/Stryker4sMavenRunner.scala
@@ -3,7 +3,7 @@ package stryker4s.maven
 import org.apache.maven.project.MavenProject
 import org.apache.maven.shared.invoker.{DefaultInvoker, Invoker}
 import stryker4s.config.Config
-import stryker4s.mutants.applymutants.ActiveMutationContext.{envVar, ActiveMutationContext}
+import stryker4s.mutants.applymutants.ActiveMutationContext.{ActiveMutationContext, envVar}
 import stryker4s.mutants.findmutants.SourceCollector
 import stryker4s.report.Reporter
 import stryker4s.maven.runner.MavenMutantRunner
